Better make sure the two files are different, so we keep the file we just uploaded

+++ packages/web/lib/pages/snapinmanagement.page.php
@@ -1715,7 +1715,9 @@ class SnapinManagement extends FOGPage
             self::$FOGSSH->put($src, $dest);
             self::$FOGSSH->disconnect();
         }
-        $this->obj->deleteFile();
+        if ($snapinfile != $this->obj->get('file')) {
+            $this->obj->deleteFile();
+        }
         $this->obj
             ->set('name', $snapin)
             ->set('description', $description)
